BeregningRoute har nå sin egen testfil
