+++ web/src/test/kotlin/no/nav/su/se/bakover/web/routes/behandling/BehandlingRoutesKtTest.kt
@@ -6,7 +6,6 @@ import arrow.core.right
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.nhaarman.mockitokotlin2.mock
 import io.kotest.assertions.assertSoftly
-import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.shouldNotBe
 import io.kotest.matchers.string.shouldContain
@@ -39,8 +38,6 @@ import no.nav.su.se.bakover.domain.behandling.BehandlingFactory
 import no.nav.su.se.bakover.domain.behandling.NySøknadsbehandling
 import no.nav.su.se.bakover.domain.behandling.extractBehandlingsinformasjon
 import no.nav.su.se.bakover.domain.behandling.withAlleVilkårOppfylt
-import no.nav.su.se.bakover.domain.beregning.Sats
-import no.nav.su.se.bakover.domain.beregning.fradrag.UtenlandskInntekt
 import no.nav.su.se.bakover.domain.journal.JournalpostId
 import no.nav.su.se.bakover.domain.oppdrag.Utbetaling
 import no.nav.su.se.bakover.domain.oppdrag.Utbetalingsrequest
@@ -62,8 +59,6 @@ import no.nav.su.se.bakover.web.stubs.asBearerToken
 import no.nav.su.se.bakover.web.testSusebakover
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
-import java.time.LocalDate
-import java.time.Month
 import java.util.UUID
 
 internal class BehandlingRoutesKtTest {
@@ -214,260 +209,6 @@ internal class BehandlingRoutesKtTest {
     }
 
     @Test
-    fun `opprette beregning for behandling`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            val fraOgMed = LocalDate.of(2021, Month.JANUARY, 1)
-            val tilOgMed = LocalDate.of(2021, Month.DECEMBER, 31)
-            val sats = Sats.HØY
-
-            services.behandling.oppdaterBehandlingsinformasjon(
-                objects.nySøknadsbehandling.id,
-                saksbehandler,
-                extractBehandlingsinformasjon(objects.behandling).withAlleVilkårOppfylt()
-            )
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    """
-                    {
-                        "fraOgMed":"$fraOgMed",
-                        "tilOgMed":"$tilOgMed",
-                        "sats":"${sats.name}",
-                        "fradrag":[]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.Created
-                val behandlingJson = deserialize<BehandlingJson>(response.content!!)
-                behandlingJson.beregning!!.fraOgMed shouldBe fraOgMed.toString()
-                behandlingJson.beregning.tilOgMed shouldBe tilOgMed.toString()
-                behandlingJson.beregning.sats shouldBe Sats.HØY.name
-                behandlingJson.beregning.månedsberegninger shouldHaveSize 12
-            }
-        }
-    }
-
-    @Test
-    fun `Fradrag med utenlandskInntekt oppretter beregning`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            val fraOgMed = LocalDate.of(2021, Month.JANUARY, 1)
-            val tilOgMed = LocalDate.of(2021, Month.DECEMBER, 31)
-            val sats = Sats.HØY
-
-            services.behandling.oppdaterBehandlingsinformasjon(
-                objects.nySøknadsbehandling.id,
-                saksbehandler,
-                extractBehandlingsinformasjon(objects.behandling).withAlleVilkårOppfylt()
-            )
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    //language=JSON
-                    """
-                    {
-                       "fraOgMed":"$fraOgMed",
-                       "tilOgMed":"$tilOgMed",
-                       "sats":"${sats.name}",
-                       "fradrag":[{
-                         "type":"Arbeidsinntekt",
-                         "beløp":200,
-                         "utenlandskInntekt":{
-                            "beløpIUtenlandskValuta":200,
-                            "valuta":"euro",
-                            "kurs":0.5
-                         },
-                         "periode" : {
-                            "fraOgMed":"$fraOgMed",
-                            "tilOgMed":"$tilOgMed"
-                         },
-                         "tilhører": "BRUKER"
-                      }]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.Created
-                val behandlingJson = deserialize<BehandlingJson>(response.content!!)
-                behandlingJson.beregning!!.fraOgMed shouldBe fraOgMed.toString()
-                behandlingJson.beregning.tilOgMed shouldBe tilOgMed.toString()
-                behandlingJson.beregning.sats shouldBe Sats.HØY.name
-                behandlingJson.beregning.månedsberegninger shouldHaveSize 12
-                behandlingJson.beregning.fradrag shouldHaveSize 2 // input + 1 because of forventet inntekt
-                behandlingJson.beregning.fradrag.filter { it.type == "Arbeidsinntekt" }.all {
-                    it.utenlandskInntekt == UtenlandskInntekt(
-                        beløpIUtenlandskValuta = 200,
-                        valuta = "euro",
-                        kurs = 0.5
-                    )
-                }
-            }
-        }
-    }
-
-    @Test
-    fun `Fradrag med utenlandskInntekt er null oppretter beregning`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            val fraOgMed = LocalDate.of(2021, Month.JANUARY, 1)
-            val tilOgMed = LocalDate.of(2021, Month.DECEMBER, 31)
-            val sats = Sats.HØY
-
-            services.behandling.oppdaterBehandlingsinformasjon(
-                objects.nySøknadsbehandling.id,
-                saksbehandler,
-                extractBehandlingsinformasjon(objects.behandling).withAlleVilkårOppfylt()
-            )
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    //language=JSON
-                    """
-                    {
-                        "fraOgMed":"$fraOgMed",
-                        "tilOgMed":"$tilOgMed",
-                        "sats":"${sats.name}",
-                        "fradrag": [{
-                            "type": "Arbeidsinntekt",
-                            "beløp": 200,
-                            "utenlandskInntekt": null,
-                            "periode" : {
-                                "fraOgMed":"$fraOgMed",
-                                "tilOgMed":"$tilOgMed"
-                            },
-                            "tilhører": "BRUKER"
-                        }]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.Created
-                val behandlingJson = deserialize<BehandlingJson>(response.content!!)
-                behandlingJson.beregning!!.fraOgMed shouldBe fraOgMed.toString()
-                behandlingJson.beregning.tilOgMed shouldBe tilOgMed.toString()
-                behandlingJson.beregning.sats shouldBe Sats.HØY.name
-                behandlingJson.beregning.månedsberegninger shouldHaveSize 12
-                behandlingJson.beregning.fradrag shouldHaveSize 2 // input + 1 because of forventet inntekt
-                behandlingJson.beregning.fradrag.all { it.utenlandskInntekt == null }
-            }
-        }
-    }
-
-    @Test
-    fun `beregn error handling`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/blabla/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {}.apply {
-                assertSoftly {
-                    response.status() shouldBe HttpStatusCode.BadRequest
-                    response.content shouldContain "ikke en gyldig UUID"
-                }
-            }
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${UUID.randomUUID()}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ).apply {
-                assertSoftly {
-                    response.status() shouldBe HttpStatusCode.BadRequest
-                    response.content shouldContain "Ugyldig body"
-                }
-            }
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ).apply {
-                assertSoftly {
-                    response.status() shouldBe HttpStatusCode.BadRequest
-                    response.content shouldContain "Ugyldig body"
-                }
-            }
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    """
-                    {
-                        "fraOgMed":"${LocalDate.of(2021, Month.JANUARY, 16)}",
-                        "tilOgMed":"${LocalDate.of(2021, Month.DECEMBER, 31)}",
-                        "sats":"ORDINÆR",
-                        "fradrag":[]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.BadRequest
-                response.content shouldContain "Ugyldige input-parametere"
-            }
-            val fraOgMed = LocalDate.of(2021, Month.JANUARY, 1)
-            val tilOgMed = LocalDate.of(2021, Month.DECEMBER, 31)
-            val sats = Sats.HØY
-
-            objects.behandling.oppdaterBehandlingsinformasjon(
-                saksbehandler,
-                extractBehandlingsinformasjon(objects.behandling).withAlleVilkårOppfylt()
-            )
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    //language=JSON
-                    """{
-                           "fraOgMed":"$fraOgMed",
-                           "tilOgMed":"$tilOgMed",
-                           "sats":"${sats.name}",
-                           "fradrag":[{
-                             "type":"Arbeidsinntekt",
-                             "beløp":200,
-                             "utenlandskInntekt":{
-                                "beløpIUtenlandskValuta":-200,
-                                "valuta":"euro",
-                                "kurs":0.5
-                             },
-                             "tilhører": "BRUKER"
-                          }]
-                        }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.BadRequest
-            }
-        }
-    }
-
-    @Test
     fun simulering() {
         withTestApplication({
             testSusebakover()
@@ -513,76 +254,6 @@ internal class BehandlingRoutesKtTest {
                 listOf(Brukerrolle.Saksbehandler)
             ) {}.apply {
                 response.status() shouldBe HttpStatusCode.OK
-            }
-        }
-    }
-
-    @Test
-    fun `client notified about illegal operations on current state of behandling`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            objects.behandling.status() shouldBe Behandling.BehandlingsStatus.OPPRETTET
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    """
-                    {
-                        "fraOgMed":"${1.januar(2021)}",
-                        "tilOgMed":"${31.desember(2021)}",
-                        "sats":"${Sats.HØY}",
-                        "fradrag":[]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.BadRequest
-                response.content shouldContain "Illegal operation"
-                response.content shouldContain "opprettBeregning"
-                response.content shouldContain "state: OPPRETTET"
-            }
-        }
-    }
-
-    @Test
-    fun `ikke lov å beregne tidligere enn 2021`() {
-        withTestApplication({
-            testSusebakover()
-        }) {
-            val objects = setup()
-            val fraOgMed = LocalDate.of(2020, Month.JANUARY, 1)
-            val tilOgMed = LocalDate.of(2020, Month.DECEMBER, 31)
-            val sats = Sats.HØY
-
-            services.behandling.oppdaterBehandlingsinformasjon(
-                objects.nySøknadsbehandling.id,
-                saksbehandler,
-                extractBehandlingsinformasjon(objects.behandling).withAlleVilkårOppfylt()
-            )
-
-            defaultRequest(
-                HttpMethod.Post,
-                "$sakPath/${objects.sak.id}/behandlinger/${objects.nySøknadsbehandling.id}/beregn",
-                listOf(Brukerrolle.Saksbehandler)
-            ) {
-                setBody(
-                    """
-                    {
-                        "fraOgMed":"$fraOgMed",
-                        "tilOgMed":"$tilOgMed",
-                        "sats":"${sats.name}",
-                        "fradrag":[]
-                    }
-                    """.trimIndent()
-                )
-            }.apply {
-                response.status() shouldBe HttpStatusCode.BadRequest
-                response.content shouldContain "En stønadsperiode kan ikke starte før 2021"
             }
         }
     }
